backend/requirements.txt
fastapi
uvicorn
authx
pydantic
python-dotenv
passlib[bcrypt]
sqlalchemy[asyncio]
asyncpg
alembic

